ajustes, mejoras y correcciones en migraciones

+++ modules/Accounting/Database/Migrations/2019_05_16_135019_add_field_institution_departament_to_accounting_seats_table.php
@@ -31,8 +31,10 @@ class AddFieldInstitutionDepartamentToAccountingSeatsTable extends Migration
      */
     public function down()
     {
-        Schema::table('accounting_seats', function (Blueprint $table) {
-            $table->dropColumn(['institution_id', 'department_id']);
-        });
+        if (Schema::hasTable('accounting_seats')) {
+            Schema::table('accounting_seats', function (Blueprint $table) {
+                $table->dropColumn(['institution_id', 'department_id']);
+            });
+        }
     }
 }
