+++ modules/Accounting/Database/Migrations/2019_05_20_134551_change_fields_tot_debit_tot_assets_to_accounting_seats_table.php
@@ -26,8 +26,10 @@ class ChangeFieldsTotDebitTotAssetsToAccountingSeatsTable extends Migration
      */
     public function down()
     {
-        Schema::table('accounting_seats', function (Blueprint $table) {
-            $table->dropColumn(['tot_debit', 'tot_assets']);
-        });
+        if (Schema::hasTable('accounting_seats')) {
+            Schema::table('accounting_seats', function (Blueprint $table) {
+                $table->dropColumn(['tot_debit', 'tot_assets']);
+            });
+        }
     }
 }
