+++ modules/Accounting/Database/Migrations/2019_05_20_144135_change_fields_debit_and_assets_to_accounting_seat_accounts_table.php
@@ -30,8 +30,10 @@ class ChangeFieldsDebitAndAssetsToAccountingSeatAccountsTable extends Migration
      */
     public function down()
     {
-        Schema::table('accounting_seat_accounts', function (Blueprint $table) {
-            $table->dropColumn(['debit', 'assets']);
-        });
+        if (Schema::hasTable('accounting_seat_accounts')) {
+            Schema::table('accounting_seat_accounts', function (Blueprint $table) {
+                $table->dropColumn(['debit', 'assets']);
+            });
+        }
     }
 }
